EFM32JGB1B: GCC: introduce efm32jg1b200f128gm48-boot-ram linker script
Example of a linker script for a boot appliaction running from RAM.

The linker script is based on efm32jg1b.ld and has been reworked:
- Provide 32 KB of FLASH memory
- Provide 30 KB of RAM_C (code) memory with cache support
- Provide 2 KB of RAM (data)

The RAM_C is mapped to the physical RAM. So, RAM_C + RAM can not
exceed the total RAM capabilities of the device.

The difference between RAM_C and RAM is, that is supports execution
with cache support.

Only the vector table, startup code and some tables required by the
startup are placed in the FLASH to be executed.

The rest of the program will be copied from FLASH to RAM_C during
the startup.

RAMFUNC functions (section .ram) are part of .text.ram.

It requires to enable __STARTUP_COPY_MULTIPLE for the startup code,
and optional __STARTUP_CLEAR_RAM_MULTIPLE to enable the support to
clear the RAM regions at startup.

Signed-off-by: Alexander Herfurtner <mission@alexanderherfurtner.de>